+++ test/twg2/text/test/StringCompareTest.java
@@ -8,8 +8,8 @@ import java.util.Map.Entry;
 import org.junit.Assert;
 import org.junit.Test;
 
+import twg2.junitassist.checks.CheckTask;
 import twg2.text.stringSearch.StringCompare;
-import checks.CheckTask;
 
 /**
  * @author TeamworkGuy2
@@ -21,10 +21,10 @@ public class StringCompareTest {
 	public void anyEndsWith() {
 		List<String> strs = Arrays.asList("String A", "String B", "Thing C");
 
-		Assert.assertTrue(StringCompare.anyStartWith(strs, new StringBuilder("String")));
-		Assert.assertTrue(StringCompare.anyStartWith(strs, new StringBuilder("Thing")));
-		Assert.assertFalse(StringCompare.anyStartWith(strs, new StringBuilder("ing")));
-		Assert.assertTrue(StringCompare.anyStartWith(strs, new StringBuilder("A=String"), 2));
+		Assert.assertTrue(StringCompare.anyStartWith(strs, charSeq("String")));
+		Assert.assertTrue(StringCompare.anyStartWith(strs, charSeq("Thing")));
+		Assert.assertFalse(StringCompare.anyStartWith(strs, charSeq("ing")));
+		Assert.assertTrue(StringCompare.anyStartWith(strs, charSeq("A=String"), 2));
 	}
 
 
@@ -65,7 +65,7 @@ public class StringCompareTest {
 
 
 	@Test
-	public void testContains() {
+	public void contains() {
 		String[] strs = { "alpha", "beta", "gamma", "int", "Integer", "IntList" };
 		String[] contains = { "gamma", "nt", "Int", "IS" };
 		Boolean[] expectEqualIgnoreCase = { true, false, true, false };
@@ -122,9 +122,11 @@ public class StringCompareTest {
 
 	@Test
 	public void compareStartsWith() {
-		Assert.assertTrue(0 == StringCompare.compareStartsWith("Java", new StringBuilder("J"), 0));
-		Assert.assertTrue(0 > StringCompare.compareStartsWith("Java", new StringBuilder("-Jazz"), 1));
-		Assert.assertTrue(0 < StringCompare.compareStartsWith("Java", new StringBuilder("==Jab"), 2));
+		Assert.assertTrue(0 > StringCompare.compareStartsWith("", charSeq("J"), 0));
+		Assert.assertTrue(0 > StringCompare.compareStartsWith("J", charSeq("Ja"), 0));
+		Assert.assertTrue(0 == StringCompare.compareStartsWith("Java", charSeq("J"), 0));
+		Assert.assertTrue(0 > StringCompare.compareStartsWith("Java", charSeq("-Jazz"), 1));
+		Assert.assertTrue(0 < StringCompare.compareStartsWith("Java", charSeq("==Jab"), 2));
 	}
 
 
@@ -139,33 +141,75 @@ public class StringCompareTest {
 
 	@Test
 	public void startsWith() {
-		Assert.assertFalse(StringCompare.startsWith("GT450".toCharArray(), 0, "A".toCharArray(), 0));
-		Assert.assertFalse(StringCompare.startsWith("GT450".toCharArray(), 0, "GA".toCharArray(), 1));
+		Assert.assertFalse(StringCompare.startsWith(charAry("GT450"), 0, charAry("A"), 0));
+		Assert.assertFalse(StringCompare.startsWith(charAry("GT450"), 0, charAry("GA"), 1));
 
-		Assert.assertTrue(StringCompare.startsWith("GT450".toCharArray(), 0, "GT".toCharArray(), 0));
-		Assert.assertTrue(StringCompare.startsWith("GT450".toCharArray(), 0, "GT450".toCharArray(), 0));
-		Assert.assertTrue(StringCompare.startsWith("-GT450".toCharArray(), 1, "==GT".toCharArray(), 2));
+		Assert.assertTrue(StringCompare.startsWith(charAry("GT450"), 0, charAry("GT"), 0));
+		Assert.assertTrue(StringCompare.startsWith(charAry("GT450"), 0, charAry("GT450"), 0));
+		Assert.assertTrue(StringCompare.startsWith(charAry("-GT450"), 1, charAry("==GT"), 2));
+	}
+
+
+	@Test
+	public void containsAll() {
+		Assert.assertFalse(StringCompare.containsAll("My Test String", new String[] { }));
+		Assert.assertFalse(StringCompare.containsAll("My Test String", new String[] { "z" }));
+		Assert.assertFalse(StringCompare.containsAll("My Test String", new String[] { "ing", "Bb" }));
+
+		Assert.assertTrue(StringCompare.containsAll("My Test String", new String[] { "" }));
+		Assert.assertTrue(StringCompare.containsAll("My Test String", new String[] { "ing", " " }));
+		Assert.assertTrue(StringCompare.containsAll("My Test String", new String[] { "Test", "My Test String", " " }));
+
+		Assert.assertFalse(StringCompare.containsAll("My Test String", Arrays.asList()));
+		Assert.assertFalse(StringCompare.containsAll("My Test String", Arrays.asList("z")));
+		Assert.assertFalse(StringCompare.containsAll("My Test String", Arrays.asList("ing", "Bb")));
+
+		Assert.assertTrue(StringCompare.containsAll("My Test String", Arrays.asList("")));
+		Assert.assertTrue(StringCompare.containsAll("My Test String", Arrays.asList("ing", " ")));
+		Assert.assertTrue(StringCompare.containsAll("My Test String", Arrays.asList("Test", "My Test String", " ")));
 	}
 
 
 	@Test
 	public void containsAny() {
 		Assert.assertFalse(StringCompare.containsAny("My Test String", new String[] { }));
+		Assert.assertFalse(StringCompare.containsAny("My Test String", new String[] { "z" }));
 		Assert.assertFalse(StringCompare.containsAny("My Test String", new String[] { "Abc", "B" }));
 
+		Assert.assertTrue(StringCompare.containsAny("My Test String", new String[] { "" }));
 		Assert.assertTrue(StringCompare.containsAny("My Test String", new String[] { "Thing", " " }));
 		Assert.assertTrue(StringCompare.containsAny("My Test String", new String[] { "X", "My Test String", " " }));
+
+		Assert.assertFalse(StringCompare.containsAny("My Test String", Arrays.asList()));
+		Assert.assertFalse(StringCompare.containsAny("My Test String", Arrays.asList("z")));
+		Assert.assertFalse(StringCompare.containsAny("My Test String", Arrays.asList("Abc", "B")));
+
+		Assert.assertTrue(StringCompare.containsAny("My Test String", Arrays.asList("")));
+		Assert.assertTrue(StringCompare.containsAny("My Test String", Arrays.asList("Thing", " ")));
+		Assert.assertTrue(StringCompare.containsAny("My Test String", Arrays.asList("X", "My Test String", " ")));
 	}
 
 
 	@Test
 	public void equal() {
-		Assert.assertFalse(StringCompare.equal("My Test String", new StringBuilder("Abc")));
-		Assert.assertFalse(StringCompare.equal("My Test String".toCharArray(), 0, "Abc".toCharArray(), 0, 3));
+		Assert.assertFalse(StringCompare.equal("My Test String", 0, "My Test", 0, 14));
+		Assert.assertFalse(StringCompare.equal("My Test String", charSeq("Abc")));
+		Assert.assertFalse(StringCompare.equal(charAry("My Test String"), 0, charAry("Abc"), 0, 3));
 
-		Assert.assertTrue(StringCompare.equal("My Test String", new StringBuilder("My Test String")));
-		Assert.assertTrue(StringCompare.equal("My Test String".toCharArray(), 0, "My ".toCharArray(), 0, 3));
-		Assert.assertTrue(StringCompare.equal("My Test String".toCharArray(), 2, "Your Test".toCharArray(), 4, 5));
+		Assert.assertTrue(StringCompare.equal("My Test String", 0, "My Test String", 0, 14));
+		Assert.assertTrue(StringCompare.equal("My Test String", charSeq("My Test String")));
+		Assert.assertTrue(StringCompare.equal(charAry("My Test String"), 0, charAry("My "), 0, 3));
+		Assert.assertTrue(StringCompare.equal(charAry("My Test String"), 2, charAry("Your Test"), 4, 5));
+	}
+
+
+	private static char[] charAry(String str) {
+		return str.toCharArray();
+	}
+
+
+	private static CharSequence charSeq(String str) {
+		return new StringBuilder(str);
 	}