+++ test/twg2/text/test/StringSplitTest.java
@@ -9,8 +9,8 @@ import java.util.Map;
 import org.junit.Assert;
 import org.junit.Test;
 
+import twg2.junitassist.checks.CheckTask;
 import twg2.text.stringUtils.StringSplit;
-import checks.CheckTask;
 
 /**
  * @author TeamworkGuy2